# commons/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "IOTGlue NodeAPI's",
    "description": "",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/iotglueapi/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
  "/commune/sources": {
      "post": {
        "description": "",
        "summary": "createSource",
        "tags": [
          "Source"
        ],
        "operationId": "createSource",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createSourceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "",
        "summary": "getAllSourcesWithPagination",
        "tags": [
          "Source"
        ],
        "operationId": "getAllSourcesWithPagination",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "page_no",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "page_size",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/common/rightpanel/flows/{flowId}": {
      "get": {
        "description": "getFlowForRightPanelById",
        "summary": "getFlowForRightPanelById",
        "tags": [
          "Common - Right Panel"
        ],
        "operationId": "getFlowForRightPanelById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [{
            "name": "flowId",
            "in": "path",
            "description": "ID of Flow",
            "required": true,
            "type": "string"
            }],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/common/sources/filter": {
      "get": {
        "summary": "getFilteredSources",
        "tags": [
          "Common"
        ],
        "operationId": "getFilteredSources",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "mappingReq",
            "in": "query",
            "required": true,
            "type": "boolean",
            "description": ""
          },
          {
            "name": "protocolType",
            "in": "query",
            "required": false,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/common/flows/filter": {
      "get": {
        "description": "get Filtered flows for Vnode",
        "summary": "getFilteredFlows",
        "tags": [
          "RTBuilder - Common"
        ],
        "operationId": "getFilteredFlows",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "isVnodeDetails",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "flowId",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/common/flows/source/link": {
      "post": {
        "description": "connect source in Commune to Vnode flow",
        "summary": "linkToCommune",
        "tags": [
          "RTBuilder - Common"
        ],
        "operationId": "linkToCommune",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/linkToCommune"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/rtbuilder/flows": {
      "post": {
        "description": "",
        "summary": "createFlow",
        "tags": [
          "Flow"
        ],
        "operationId": "createFlow",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createFlowRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/xbuilder/dashboards": {
      "post": {
        "description": "",
        "summary": "createDashboard",
        "tags": [
          "Dashboard"
        ],
        "operationId": "createdashboard",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createDashboardRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/xbuilder/dashboards/{dashboardId}": {
      "get": {
        "description": "",
        "summary": "getDashboardById",
        "tags": [
          "Dashboard"
        ],
        "operationId": "getDashboardById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "dashboardId",
            "in": "path",
            "description": "ID of dashboard to return",
            "required": true,
            "type": "string"
            }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    }
},
  "definitions": {
    "linkToCommune": {
      "title": "linkToCommune",
      "example": {
        "sourceData": {
          "name": "flow some",
          "protocolDetails": {
            "protocol": "HTTPS",
            "type": "https",
            "ipOrDomain": "172.30.66.175",
            "port": "8089"
          },
          "url": "https://172.30.66.175:8089",
          "templateId": "",
          "sourceMappingReqd": false,
          "createdBy": "Ajay Ramesh",
          "resourceDetails": [
            {
              "id": "SG_RD_1",
              "name": "asdasd",
              "description": "asdasd",
              "methodOrType": "POST",
              "header": [
                "Content-Type:application/json"
              ],
              "uniqueKey": "",
              "isInterpolated": false,
              "uriOrTopic": "/vnode-application/rest/viewnode/rulebreak/SG_2/FL_34_VN_2",
              "pollingTime": 10,
              "paramType": "noparam",
              "requestPayload": {
                "fileName": "",
                "payload": {
                  "isSourceBasedFlow": true,
                  "sourceMappingId": [
                    "SM_2",
                    "SM_1"
                  ],
                  "endpointSelected": [
                    3,
                    4,
                    1,
                    2
                  ],
                  "uniquekey": "deviceId"
                }
              },
              "responsePayload": {
                "fileName": "",
                "payload": [
                  {
                    "deviceId": 34,
                    "temp": 45,
                    "hum": 67,
                    "iotglueentrytime": 1600159565477
                  }
                ]
              },
              "dateTime": [
                {
                  "key": "",
                  "format": ""
                }
              ],
              "date": [
                {
                  "key": "",
                  "format": ""
                }
              ],
              "time": [
                {
                  "key": "",
                  "format": ""
                }
              ]
            }
          ],
          "createdOn": "",
          "isDeleted": false
        },
        "linkData": {
          "flowId": "FL_100",
          "isVnodeLinked": true
        }
      },
      "type": "object",
      "properties": {
        "sourceData": {
          "$ref": "#/definitions/SourceData"
        },
        "linkData": {
          "$ref": "#/definitions/LinkData"
        }
      },
      "required": [
        "sourceData",
        "linkData"
      ]
    },
...
    "createFlowRequest": {
      "title": "createFlowRequest",
      "example": {
        "name": "Flow 10 Approved",
        "status": "Approved",
        "description": "Source-WithUKWithTemp",
        "nodeType": "Flow",
        "statusUpdateComment": "Right flow",
        "permissionedBy": "Archana",
        "nodes": {
          "Source": {
            "Source_0": {
              "nodeType": "SourceGroup",
              "name": "SourceGroupName",
              "uniqueResourceSelectedStatus": true,
              "sourceId": "SG_5",
              "sourceMappingId": "SM_4",
              "url": "http://127.0.0.1:8023",
              "uriOrTopic": "/data?id={deviceID}",
              "protocol": "HTTP",
              "uniqueResourceSelected": [
                "12345Device"
              ],
              "methodOrType": "GET",
              "count": 0,
              "xPosition": 0.275,
              "yPosition": 0.238888888888889
            }
          },
          "Rules": {
            "Rule_1": {
              "id": "RL_7",
              "name": "TempRule",
              "nodeName": "R1",
              "description": "exp",
              "type": "simpleExp",
              "simpleExp": "(temp>27)",
              "attributeKeysDataType": {
                "temp": "String"
              },
              "nodeType": "Rule",
              "count": 1,
              "xPosition": 0.13125,
              "yPosition": 0.125555555555556
            },
            "Rule_2": {
              "id": "RL_8",
              "name": "TempRule",
              "nodeName": "R2",
              "description": "script",
              "type": "script",
              "script": "RTB_SC_4",
              "attributeKeysDataType": {
                "temp": "Integer",
                "deviceID": "String"
              },
              "nodeType": "Rule",
              "count": 1,
              "xPosition": 0.13125,
              "yPosition": 0.125555555555556
            }
          },
          "Actions": {
            "Action_2": {
              "id": "RTB_AC_01",
              "name": "EMAIL",
              "description": "Send mail",
              "typeOfAction": "email",
              "fields": {
                "cc": "swathi_somashekar@thbs.com,prashannakumar_somasundaram@thbs.com,krishnapriya_varanasi@thbs.com",
                "message": "Welcome to RTBuilder.Please take immediate Action since the set rule has broken.",
                "subject": "Alert!.Please take immediate Action",
                "bcc": "krishnapriya_varanasi@thbs.com",
                "to": "krishnapriya_varanasi@thbs.com",
                "signature": "Regards,\nIotGlue",
                "actionMapping": "supervisoremail, vendoremail"
              },
              "nodeType": "Action",
              "nodeName": "email",
              "count": 3,
              "xPosition": 0.13125,
              "yPosition": 0.125555555555556
            },
            "Action_4": {
              "typeOfAction": "sms",
              "name": "SMS",
              "nodeName": "R1",
              "description": "Sendmessage",
              "id": "RTB_AC_02",
              "fields": {
                "toPhoneNumber": "+919551352836,+971989466730",
                "message": "Welcometo RTBuilder. Please take immediate Action.",
                "actionMapping": "buheadnumber, managernumber"
              },
              "nodeType": "Action",
              "count": 4,
              "xPosition": 0.13125,
              "yPosition": 0.125555555555556
            },
            "Action_8": {
              "typeOfAction": "api",
              "name": "service",
              "description": "API or Aervice",
              "id": "RTB_AC_03",
              "fields": {
                "url": "http://182.22.22.2/data",
                "methodtype": "get",
                "requestbody": "",
                "responsebody": "",
                "header": {
                  "content-type": "Application/json"
                }
              },
              "nodeType": "Service",
              "count": 4,
              "xPosition": 0.13125,
              "yPosition": 0.125555555555556
            }
          },
          "Visualisation": {},
          "Trigger": {
            "Trigger_4": {
              "windowTimeValueUnit": "MINUTES",
              "windowTimeValue": "1",
              "nodeType": "Trigger",
              "endpointTriggerValueStatus": false,
              "isSourceBased": false,
              "triggerCount": [
                {
                  "triggerValue": null,
                  "uniqueKeyValue": "ZYME-9001"
                },
                {
                  "triggerValue": null,
                  "uniqueKeyValue": "ZYME-9002"
                },
                {
                  "triggerValue": null,
                  "uniqueKeyValue": "ZYME-9003"
                }
              ],
              "commonTriggerValue": 2,
              "count": 1,
              "xPosition": 0.13125,
              "yPosition": 0.125555555555556
            }
          },
          "Forward": {
            "Forward_2": {
              "timerUnit": "MINUTES",
              "timerValue": 5,
              "isSourceBased": false,
              "nodeType": "Forwarder",
              "count": 1,
              "xPosition": 0.13125,
              "yPosition": 0.125555555555556
            }
          }
        },
        "linearExpression": {
          "LX1": "Source_0->Rule_1->Action_X->end_0",
          "LX2": "Source_0->Rule_2->Action_X->end_0"
        },
        "connections": []
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "nodeType": {
          "type": "string"
        },
        "statusUpdateComment": {
          "type": "string"
        },
        "permissionedBy": {
          "type": "string"
        },
        "nodes": {
          "$ref": "#/definitions/Nodes"
        },
        "linearExpression": {
          "$ref": "#/definitions/LinearExpression"
        },
        "connections": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Connection"
          }
        }
      },
      "required": [
        "name",
        "status",
        "description",
        "nodeType",
        "statusUpdateComment",
        "permissionedBy",
        "nodes",
        "linearExpression",
        "connections"
      ]
    },
...
    "LinearExpression": {
      "title": "LinearExpression",
      "example": {
        "LX1": "Source_0->Rule_1->Action_X->end_0",
        "LX2": "Source_0->Rule_2->Action_X->end_0"
      },
      "type": "object",
      "properties": {
        "LX1": {
          "type": "string"
        },
        "LX2": {
          "type": "string"
        }
      },
      "required": [
        "LX1",
        "LX2"
      ]
    },
...
    "createSourceRequest": {
      "title": "createSourceRequest",
      "example": {
        "name": "Source",
        "protocolDetails": {
          "protocol": "MQTT",
          "type": "tcp",
          "ipOrDomain": "192.168.0.1",
          "port": 1883
        },
        "url": "tcp://192.168.0.1:1883",
        "templateId": "1",
        "sourceMappingReqd": true,
        "resourceDetails": [
          {
            "id": "SG_RD_1",
            "name": "First Endpoint",
            "description": "Source with Unique key and  with template",
            "methodOrType": "GET",
            "header": {
              "content-type": "Application/json"
            },
            "responsePayload": {
              "fileName": "",
              "payload": {
                "device_id": 75,
                "temp": 30
              },
              "schema": {
                "type": "object",
                "properties": {
                  "device_id": {
                    "type": "number"
                  },
                  "temp": {
                    "type": "number"
                  }
                },
                "required": [
                  "device_id",
                  "temp",
                  "humd"
                ]
              }
            },
            "requestPayload": {},
            "uniqueKey": "device_id",
            "uriOrTopic": "topic-1",
            "pollingTime": 0,
            "paramType": "none",
            "dateTime": {}
          }
        ],
        "createdBy": "Santosh"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "protocolDetails": {
          "$ref": "#/definitions/ProtocolDetails"
        },
        "url": {
          "type": "string"
        },
        "templateId": {
          "type": "string"
        },
        "sourceMappingReqd": {
          "type": "boolean"
        },
        "resourceDetails": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ResourceDetail"
          }
        },
        "createdBy": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "protocolDetails",
        "url",
        "templateId",
        "sourceMappingReqd",
        "resourceDetails",
        "createdBy"
      ]
    },
    "ProtocolDetails": {
      "title": "ProtocolDetails",
      "example": {
        "protocol": "MQTT",
        "type": "tcp",
        "ipOrDomain": "192.168.0.1",
        "port": 1883
      },
      "type": "object",
      "properties": {
        "protocol": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "ipOrDomain": {
          "type": "string"
        },
        "port": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "protocol",
        "type",
        "ipOrDomain",
        "port"
      ]
    },
    "ResourceDetail": {
      "title": "ResourceDetail",
      "example": {
        "id": "SG_RD_1",
        "name": "First Endpoint",
        "description": "Source with Unique key and  with template",
        "methodOrType": "GET",
        "header": {
          "content-type": "Application/json"
        },
        "responsePayload": {
          "fileName": "",
          "payload": {
            "device_id": 75,
            "temp": 30
          },
          "schema": {
            "type": "object",
            "properties": {
              "device_id": {
                "type": "number"
              },
              "temp": {
                "type": "number"
              }
            },
            "required": [
              "device_id",
              "temp",
              "humd"
            ]
          }
        },
        "requestPayload": {},
        "uniqueKey": "device_id",
        "uriOrTopic": "topic-1",
        "pollingTime": 0,
        "paramType": "none",
        "dateTime": {}
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "methodOrType": {
          "type": "string"
        },
        "header": {
          "$ref": "#/definitions/Header"
        },
        "responsePayload": {
          "$ref": "#/definitions/ResponsePayload"
        },
        "requestPayload": {
          "type": "object"
        },
        "uniqueKey": {
          "type": "string"
        },
        "uriOrTopic": {
          "type": "string"
        },
        "pollingTime": {
          "type": "integer",
          "format": "int32"
        },
        "paramType": {
          "type": "string"
        },
        "dateTime": {
          "type": "object"
        }
      },
      "required": [
        "id",
        "name",
        "description",
        "methodOrType",
        "header",
        "responsePayload",
        "requestPayload",
        "uniqueKey",
        "uriOrTopic",
        "pollingTime",
        "paramType",
        "dateTime"
      ]
    },
...
  },
...
}